requirements.txt
streamlit>=1.30.0
ruamel.yaml>=0.18.0
requests>=2.31.0
python-dotenv>=1.0.0
